// src/interfaces/madeira.interface.ts
export interface Madeira {
    id?: string | null | undefined;
    data?: string;
    placa?: string;
    processada?: string;
    nomePopular?: string;
    nomeCientifico?: string;
    tora?: string;
    produto?: string;
    expessura?: number;
    largura?: number;
    comprimento?: number;
    quantidade?: number;
    metroCubico?: number;
    COD?: number;
    NF?: string;
    DOF?: string;
    observacao?: string;
    valor?: number;
    ativo?: boolean;
    createdAt?: string;
    updatedAt?: string;
    deletedAt?: string;
}